requirements.txt
# Python dependencies for crypto trading engine

# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.1

# Database connectivity
mysql-connector-python==8.1.0
pymysql==1.1.0
sqlalchemy==2.0.19

# API and HTTP
requests==2.31.0
aiohttp==3.8.5
fastapi==0.101.1
uvicorn==0.23.2
websockets==11.0.3

# Kubernetes and containerization
kubernetes==27.2.0

# Trading and financial
python-binance==1.0.19
ccxt==4.0.45
ta-lib==0.4.25
yfinance==0.2.20

# Machine learning and AI
torch==2.0.1
transformers==4.32.1
ollama==0.1.7

# Data processing
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Configuration
python-dotenv==1.0.0
pydantic==2.1.1
pyyaml==6.0.1

# Cryptography and security
cryptography==41.0.3
pyjwt==2.8.0
ecdsa==0.18.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
requests-mock==1.11.0

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.3.3

# Numerical computing
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
click==8.1.6
rich==13.5.2
tqdm==4.65.0